# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraderLoading API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface UserProfile {
  id: number;
  name: string;
  avatarUrl?: string | null;
  xp: number;
  level: number;
  xpToNextLevel: number;
}

export interface UpdateProfileRequest {
  name: string;
  avatarUrl?: string | null;
}

export interface Mission {
  id: number;
  title: string;
  description: string;
  xpReward: number;
  completed: boolean;
  completedAt?: string | null;
}

export interface CompleteMissionResponse {
  mission: Mission;
  profile: UserProfile;
  levelUp: boolean;
}

export interface JournalImage {
  id: number;
  url: string;
}

export type JournalEntryResult =
  (typeof JournalEntryResult)[keyof typeof JournalEntryResult];

export const JournalEntryResult = {
  win: "win",
  loss: "loss",
  breakeven: "breakeven",
  none: "none",
} as const;

export interface JournalEntry {
  id: number;
  title: string;
  content: string;
  tradeDate: string;
  result: JournalEntryResult;
  tags?: string | null;
  images: JournalImage[];
  createdAt: string;
  updatedAt: string;
}

export type CreateJournalEntryRequestResult =
  (typeof CreateJournalEntryRequestResult)[keyof typeof CreateJournalEntryRequestResult];

export const CreateJournalEntryRequestResult = {
  win: "win",
  loss: "loss",
  breakeven: "breakeven",
  none: "none",
} as const;

export interface CreateJournalEntryRequest {
  title: string;
  content: string;
  tradeDate: string;
  result: CreateJournalEntryRequestResult;
  tags?: string | null;
}

export type UpdateJournalEntryRequestResult =
  (typeof UpdateJournalEntryRequestResult)[keyof typeof UpdateJournalEntryRequestResult];

export const UpdateJournalEntryRequestResult = {
  win: "win",
  loss: "loss",
  breakeven: "breakeven",
  none: "none",
} as const;

export interface UpdateJournalEntryRequest {
  title: string;
  content: string;
  tradeDate: string;
  result: UpdateJournalEntryRequestResult;
  tags?: string | null;
}

export interface DeleteResponse {
  success: boolean;
}

export interface UploadImageResponse {
  image: JournalImage;
}

export type IdeaType = (typeof IdeaType)[keyof typeof IdeaType];

export const IdeaType = {
  idea: "idea",
  goal: "goal",
} as const;

export interface Idea {
  id: number;
  type: IdeaType;
  content: string;
  completed: boolean;
  createdAt: string;
}

export type CreateIdeaRequestType =
  (typeof CreateIdeaRequestType)[keyof typeof CreateIdeaRequestType];

export const CreateIdeaRequestType = {
  idea: "idea",
  goal: "goal",
} as const;

export interface CreateIdeaRequest {
  type: CreateIdeaRequestType;
  content: string;
}

export interface UpdateIdeaRequest {
  content: string;
  completed: boolean;
}

export interface ChecklistItem {
  id: number;
  text: string;
  completed: boolean;
  order: number;
}

export interface CreateChecklistItemRequest {
  text: string;
}

export interface UpdateChecklistItemRequest {
  text: string;
  completed: boolean;
}

export type NewsArticleSentiment =
  | (typeof NewsArticleSentiment)[keyof typeof NewsArticleSentiment]
  | null;

export const NewsArticleSentiment = {
  bullish: "bullish",
  bearish: "bearish",
  neutral: "neutral",
} as const;

export interface NewsArticle {
  title: string;
  summary: string;
  source: string;
  publishedAt?: string | null;
  url?: string | null;
  sentiment?: NewsArticleSentiment;
}

export interface NewsResponse {
  articles: NewsArticle[];
  fetchedAt: string;
  hasApiKey: boolean;
}

export type UserSettingsBackgroundType =
  (typeof UserSettingsBackgroundType)[keyof typeof UserSettingsBackgroundType];

export const UserSettingsBackgroundType = {
  default: "default",
  custom: "custom",
} as const;

export type UserSettingsFontChoice =
  (typeof UserSettingsFontChoice)[keyof typeof UserSettingsFontChoice];

export const UserSettingsFontChoice = {
  inter: "inter",
  jetbrains: "jetbrains",
  roboto: "roboto",
  "space-grotesk": "space-grotesk",
  "ibm-plex": "ibm-plex",
} as const;

export interface UserSettings {
  id: number;
  backgroundUrl?: string | null;
  backgroundType: UserSettingsBackgroundType;
  fontChoice: UserSettingsFontChoice;
  /**
   * @minimum 0
   * @maximum 90
   */
  backgroundDarkness: number;
}

export type UpdateUserSettingsRequestBackgroundType =
  (typeof UpdateUserSettingsRequestBackgroundType)[keyof typeof UpdateUserSettingsRequestBackgroundType];

export const UpdateUserSettingsRequestBackgroundType = {
  default: "default",
  custom: "custom",
} as const;

export type UpdateUserSettingsRequestFontChoice =
  (typeof UpdateUserSettingsRequestFontChoice)[keyof typeof UpdateUserSettingsRequestFontChoice];

export const UpdateUserSettingsRequestFontChoice = {
  inter: "inter",
  jetbrains: "jetbrains",
  roboto: "roboto",
  "space-grotesk": "space-grotesk",
  "ibm-plex": "ibm-plex",
} as const;

export interface UpdateUserSettingsRequest {
  backgroundUrl?: string | null;
  backgroundType?: UpdateUserSettingsRequestBackgroundType;
  fontChoice?: UpdateUserSettingsRequestFontChoice;
  /**
   * @minimum 0
   * @maximum 90
   */
  backgroundDarkness?: number;
}

export interface UploadBackgroundResponse {
  url: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type UploadJournalImageBody = {
  image: Blob;
};

export type UploadBackgroundBody = {
  image: Blob;
};