# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TraderLoading API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /missions/{id}/complete:
    post:
      operationId: completeMission
      tags: [missions]
      summary: Mark a mission as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated mission and profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteMissionResponse"
  /journal:
    get:
      operationId: getJournalEntries
      tags: [journal]
      summary: Get all journal entries
      responses:
        "200":
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a new journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalEntryRequest"
      responses:
        "201":
          description: Created journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journal/{id}:
    get:
      operationId: getJournalEntry
      tags: [journal]
      summary: Get a single journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    put:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJournalEntryRequest"
      responses:
        "200":
          description: Updated journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /journal/{id}/images:
    post:
      operationId: uploadJournalImage
      tags: [journal]
      summary: Upload an image for a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required:
                - image
      responses:
        "200":
          description: Uploaded image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadImageResponse"
  /journal/{id}/images/{imageId}:
    delete:
      operationId: deleteJournalImage
      tags: [journal]
      summary: Remove an image from a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /ideas:
    get:
      operationId: getIdeas
      tags: [ideas]
      summary: Get ideas and goals
      responses:
        "200":
          description: List of ideas and goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
    post:
      operationId: createIdea
      tags: [ideas]
      summary: Create an idea or goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIdeaRequest"
      responses:
        "201":
          description: Created idea/goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
  /ideas/{id}:
    put:
      operationId: updateIdea
      tags: [ideas]
      summary: Update an idea or goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIdeaRequest"
      responses:
        "200":
          description: Updated idea/goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
    delete:
      operationId: deleteIdea
      tags: [ideas]
      summary: Delete an idea or goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
...
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        xp:
          type: integer
        level:
          type: integer
        xpToNextLevel:
          type: integer
      required:
        - id
        - name
        - xp
        - level
        - xpToNextLevel
...
    Mission:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        xpReward:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: string
          nullable: true
      required:
        - id
        - title
        - description
        - xpReward
        - completed
    CompleteMissionResponse:
      type: object
      properties:
        mission:
          $ref: "#/components/schemas/Mission"
        profile:
          $ref: "#/components/schemas/UserProfile"
        levelUp:
          type: boolean
      required:
        - mission
        - profile
        - levelUp
    JournalImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
      required:
        - id
        - url
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        tradeDate:
          type: string
        result:
          type: string
          enum: [win, loss, breakeven, none]
        tags:
          type: string
          nullable: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/JournalImage"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - content
        - tradeDate
        - result
        - images
        - createdAt
        - updatedAt
    CreateJournalEntryRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tradeDate:
          type: string
        result:
          type: string
          enum: [win, loss, breakeven, none]
        tags:
          type: string
          nullable: true
      required:
        - title
        - content
        - tradeDate
        - result
    UpdateJournalEntryRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tradeDate:
          type: string
        result:
          type: string
          enum: [win, loss, breakeven, none]
        tags:
          type: string
          nullable: true
      required:
        - title
        - content
        - tradeDate
        - result
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    UploadImageResponse:
      type: object
      properties:
        image:
          $ref: "#/components/schemas/JournalImage"
      required:
        - image
    Idea:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [idea, goal]
        content:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - type
        - content
        - completed
        - createdAt
    CreateIdeaRequest:
      type: object
      properties:
        type:
          type: string
          enum: [idea, goal]
        content:
          type: string
      required:
        - type
        - content
    UpdateIdeaRequest:
      type: object
      properties:
        content:
          type: string
        completed:
          type: boolean
      required:
        - content
        - completed
...